Add support for checking Echo notifications
Echo is broken on ScratchWikiSkin2 "out-of-the-box" because it adds weird, broken elements into a dropdown.

Now we're filtering those and adding our own custom notifications button to the navbar. This will show if you have unread notifications, and if so, link you to Special:Notifications

In the future it would be nice to implement the pop-up instead of linking the user, and generating the link MW instead of hard-coding it, but I believe this is low-priority since it will fix itself.