Hide logo on REALLY narrow screens
The addition of the messages envelope makes the navbar overflow on
Chrome devtools "Mobile S" width. This commit hides the Scratch logo at
that point, under the assumption that the search bar, messages, pencil,
and user icons are all more important to keep.

The width value in question was chosen as the minimum width before the
navbar begins to overflow. (That value is language-independent because
there is no text in the navbar at that point.)