Add support for checking Echo notifications (#129)
* Add support for checking Echo notifications

Echo is broken on ScratchWikiSkin2 "out-of-the-box" because it adds weird, broken elements into a dropdown.

Now we're filtering those and adding our own custom notifications button to the navbar. This will show if you have unread notifications, and if so, link you to Special:Notifications

In the future it would be nice to implement the pop-up instead of linking the user, and generating the link MW instead of hard-coding it, but I believe this is low-priority since it will fix itself.

* localize and add feet class

* update feet from id to class

* new comment

* removetest

* don't show for logged-out users

* Use "messages" instead of "notifications" to be consistent with Scratch
...to the extent Scratch is self-consistent. Looking at you, nav-"notifications".png!

* Code cleanup on Mrsrec's work

* Use Scratch-style notification indicator, incl. count

* SpecialPage::getTitleFor and $user->isRegistered()

* Prefer getLinkURL() and against PHP echo

* Update CREDITS
Need to include message envelope icon, and Scratch code lol

* Don't name stylesheets based on absolute width

* box-sizing: border-box to avoid weird (literal) edge cases

* Don't hardcode desktop width in JS

* Hide logo on REALLY narrow screens

The addition of the messages envelope makes the navbar overflow on
Chrome devtools "Mobile S" width. This commit hides the Scratch logo at
that point, under the assumption that the search bar, messages, pencil,
and user icons are all more important to keep.

The width value in question was chosen as the minimum width before the
navbar begins to overflow. (That value is language-independent because
there is no text in the navbar at that point.)

* Fix ResourceLoader warning

* Use ResourceLoader logo preload

* Check if $tab['class'] exists

The reason this check was initially omitted was because I mistakenly thought it was guaranteed to exist by the code above

---------

Co-authored-by: AbyxDev <ken@abyx.dev>